LPD-57577 Remove MVCResourceCommand since it only retrieves all user groups, and simplify
