--- modules/apps/notification/notification-web/src/main/java/com/liferay/notification/web/internal/portlet/action/GetEmailNotificationUserGroupsMVCResourceCommand.java
@@ -1,65 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.notification.web.internal.portlet.action;
-
-import com.liferay.notification.constants.NotificationPortletKeys;
-import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.model.UserGroup;
-import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
-import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCResourceCommand;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
-import com.liferay.portal.kernel.service.UserGroupLocalService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import jakarta.portlet.ResourceRequest;
-import jakarta.portlet.ResourceResponse;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Christopher Kian
- */
-@Component(
-	property = {
-		"jakarta.portlet.name=" + NotificationPortletKeys.NOTIFICATION_TEMPLATES,
-		"mvc.command.name=/notification_templates/get_email_notification_user_groups"
-	},
-	service = MVCResourceCommand.class
-)
-public class GetEmailNotificationUserGroupsMVCResourceCommand
-	extends BaseMVCResourceCommand {
-
-	@Override
-	protected void doServeResource(
-			ResourceRequest resourceRequest, ResourceResponse resourceResponse)
-		throws Exception {
-
-		JSONArray userGroupJSONArray = _jsonFactory.createJSONArray();
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)resourceRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		for (UserGroup userGroup :
-				_userGroupLocalService.getUserGroups(
-					themeDisplay.getCompanyId())) {
-
-			userGroupJSONArray.put(userGroup.getName());
-		}
-
-		JSONPortletResponseUtil.writeJSON(
-			resourceRequest, resourceResponse, userGroupJSONArray);
-	}
-
-	@Reference
-	private JSONFactory _jsonFactory;
-
-	@Reference
-	private UserGroupLocalService _userGroupLocalService;
-
-}
